requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.5
requests==2.32.3
python-multipart==0.0.20
starlette==0.41.3
python-dotenv==1.0.1